[sysmem] sysmem-connector zircon service
Make fuchsia_sysmem_Allocator2 a zircon service.  The service finds a
sysmem device instance and forwards the service request to the driver
to be served directly from there.

This way a non-driver client of sysmem doesn't need to hassle with
finding the sysmem device instance etc.

A similar garnet service will be added which shares the sysmem-connector
code, so to anticipate that, sysmem-connector is a ulib.

Tested: /system/test/sys/sysmem-test
Change-Id: Ie2773eb919b67a5bc53bd0a4e8b50921e33127fc