[wlan] Stay on channel for full duration during scan
Currently, inside channel scheduler, we start timer countdown for
staying on channel before we switch channel, so the actual time we stay
on channel is actually the max channel duration in the scan request
subtracted by the channel switching time. This channel switching time is
~60ms on ralink and 40-100ms on ath10k (though the latter value doesn't
matter too much since we use hardware scanning there).

This commit changes the timer countdown to start after we have already
switched channel and reduces the max duration in the scan request to
account for what was previously used by the channel switch time.

Test: Tested `wlan scan` and ten `wlan connect` on ralink and ath10k
Change-Id: Ib6f3d71898d3a94e69a187aa03dd3c0bb089c3d5