[sessionmgr] Properly employ deferred_action.
deferred_action is meant to defer an action until the end of some
scope, but we use it merely to defer it to some time later, but
decidedly not to the time of destruction, at which time it must
not execute at all if it hasn't yet.

So we use a simple fit::function<> instead that we simply protect
against double invocation using a mutable bool flag on the capture
list. This is especially simple because the capture list already
is mutable.

Amends https://fuchsia-review.googlesource.com/c/peridot/+/240436

TESTED=/pkgfs/packages/modular_tests/0/test/run_modular_tests.sh

Change-Id: I8851a349bbd4c55312fc59c443f74f98145bd2ea