[wlan][mlme] Compute and skip body padding
Some drivers, such as Ralink, align frame bodies.
This CL changes the frame parsing to take a boolean
indicating whether the body is padded instead of a
concrete padding length.
This is more convenient to use as the length of the header
was already determined by parsing it.

WLAN-441 #comment

Test=Included

Change-Id: I656968411ce53ee411f4da89b3d4d32347822ce7