[pm publish] consistent snapshot root manifest
The prior repository code was conflating repository initialization for
modification, and new repository creation. It also supported a feature to
allow a a root manifest to be injected from an otherwise unsupported key
location. The features are removed because they have unstable behaviors under
various conditions and could lead to destruction/lost data of repositories
that are being updated.

The pre-initialization of the repository from the build is now instead
performed by the build itself. I inspected other callers of `pm publish` to
ensure they were not using the keys flag - as I did not find any, the flag is
removed entirely as it is a misfeature that is tricky to statefully manage
due to a lack of features in the underlying TUF library.

The `update` package is renamed to `repo`, as it is all about managing a
repository. This has the effect of significantly improving readability of the
code.

A fixup function is added that is run at the end of a commit that ensures
that a consistent snapshot for root.json exists. This is a specific
workaround for the in-tree setup that copies in root.json, but not a snapshot
for root.json. This fixup could be performed in the build, but would
otherwise require additional fixups being distributed to other places in the
infrastructure, so this approach is satisfactory for now. It is this specific
portion of the change that is key to solving the linked bug.

The test cases are cleaned up, with coverage added for consistent snapshots
and all produced files. Coverage is also added for blob publishing details.
The feature set coverage is now sufficient to provide reasonable confidence
of both in-tree and out-of-tree common use cases via `pm newrepo` and `pm
publish`.

Previously the code would not result in a call to repo.Init() in the case
where root_manifest.json was used to initialize a repository. If the
root_manifest.json constains consistent_snapshot: true, the resultant
repository would be missing the consistent snapshot of the root manifest
(<hash>.root.json), resulting in errors when syncing the repositories.

Various dead code is deleted.

Test: pm newrepo -k garnet/go/src/amber/keys produces a repo that includes <hash>.root.json.
Test: observe no <hash>.root.json in repositories before. observe <hash>.root.json in repositories after.
Test: CQ, for new test coverage.
Bug: INTK-781 #done
Change-Id: I26c0baf6e7d6c759645667254f9b4232a03f97eb