[scsilib] Add Controller interface between SCSI targets and HBAs
Introduce Controller, a very simple interface between SCSI target
drivers (ex: Disks) and SCSI Controller drivers. It provides a
mechanism for targets to execute commands with a data-in/data-out
region and retrieve the status of the command.

This interface is purely synchronous and non-queued. It is only in place
to support development. In the future it should be replaced with a
Banjo interface.

The plan (see conflicting WIP CL) is that peripherals will be constructed
with a pointer to their Controller and will use this interface to execute
initial (probe, startup, enumerate) commands.

ZX-2314

Tested: Built

Change-Id: I6e25bbd80a79e6dbf652c6a28b980d850ef32d4f