[scenic] Add view state tests
Add end to end view state tests.  View state tests run a full scenic
application via the |EmbedderView| utility, and verify that they
successfully produce a |ViewStateChangedEvent|.

Follow up work involves testing additional apps (for now we just test
the bouncing_ball app), and verifying more stuff (e.g., pixel test that
the screen is not black when we know the screen should be "something").
Let's keep the scope small for now though.

A big thanks to Mike and Mikael for pair collaborating on this, and
also Ross for writing the scenic pixel tests, which this test is heavily
based on.

Test: CQ
fx run-test scenic_tests -t gfx_viewstate_apptests

Change-Id: Ia12ac0b6e90b9db2c20ee411c35738540ab8dd61