[sysmem][astro] Allocate protected memory region on boot
Use metadata to specify the size of a protected memory region, which
will be created using contigous memory. Portions of that memory region
can be handed out to clients as requested. Currently the memory isn't
actually marked as protected using the TEE, so it's no different from
regular memory.

Also convert the query of the current VID and PID to read from the
sysmem metadata, as pdev_get_device_info gets information about the
current device, which always PDEV_VID_GENERIC and PDEV_PID_GENERIC for
sysmem.

Test: vkcube on astro

Change-Id: I681f928ccde4e4d99787d2b0853e19ed9b9eaab0