[images] faster incremental blob publishing
copytree was only used to produce the blob output directory. The blob output
directory is content addressed, as such repeating writes to target locations
is not useful work. The code for copytree is used to produce a copy-only-once
behavior that significantly reduces total number and cost of io operations
for incremental builds.

Change-Id: Idc4b14b88326f93cfe2ff5c43e01a6628f7b2bec