[inspect] Make Object self-ownership more obvious
Object held a BindingSet of RefPtrs to itself. This is not well-handled
by BindingSet, because BindingSet expects to be alive after after
destructing a binding. This change cuts BindingSet out of the ownership
loop: Object holds a RefPtr to itself whenever the BindingSet is
non-empty (using the empty set handler to track emptiness). By design,
it is safe to destruct the BindingSet when it calls its empty set
handler.

TEST=asan
CF-344: #done

Change-Id: Icbb6d249093f4e986e24614bb429d610945a09f3