[fidl][light] Add FIDL API for light drivers
This API will be used to support hardware lights such as indicator LEDs.
The lights are identifiable by a board specific name.
A light device may implement one or more physical lights,
allowing for an array of LEDs grouped together
(for example, the LED ring on the Nexus Q music player).
Lights may be simple on/off, support a brightness value, or an RGB value.

Bug: ZX-3349

TEST: compiles when included in astro board driver.
Change-Id: Ib1b19dd27a6cf12f7171debd4a4c8715cdbe2460