[bt][gap] fix use-after-free in some RemoteDeviceCache tests
In the GAP_RemoteDeviceCacheTest ExpirationTest test cases, we need to
be careful about the lifetime of the RemoteDevices. In particular,
when querying the RemoteDeviceCache for a RemoteDevice, we must not
use our cached pointer to said device, as a) the RemoteDevice is owned
by the RemoteDeviceCache, and b) the cache may have expired the
device.

Update the ExpirationTest test cases, to use cached values of the
device's identifier and address, both of which are cached by value,
rather than fetching the identifier or address by dereferencing a
(potentially invalid) pointer.

Note: when validating this change, I deliberately avoided running
the entire unit test suite. That's because we still have memory-use
bugs in other parts of the suite.

Bug: BT-693
Test: fx set x64 --variant=asan/bt-host-unittests && \
      fx build && fx push-package bluetooth-tests && \
      fx run-test bluetooth-tests -t bt-host-unittests -- \
        --gtest_filter="GAP_RemoteDeviceCacheTest*"

Change-Id: Id25e58b76eb00b172901febc448933adc329adf4