[fidl][cpp] Generate test for reserved words escaping
This includes a script that generates a set of FIDL libraries that try
to use names in a variety of places that identifiers can go. These
libraries are then built for C++, testing that they can in fact be
built.

There are also a bunch of C++ bindings generator changes to allow this
to actually work.

Test: built the whole tree

Change-Id: I17cc0363683c9bc50007da0268ef3e14a14ec77d