[fxl] Make 'fx run-test fxl_unittests' work again
This is a workaround for BLD-338, which causes 'fx run-test' to not work
properly if the package_name attribute is used on the package template.
The executable target is renamed, which unfortunately changes the name
used with 'fx run-host-tests' because it is also not currently possible
because the build system assumes the binary name is the same as the
target name there.

BLD-338 #workaround
IN-819 #workaround

Test: fx run-host-tests fxl_unittests_bin
Test: fx run-test fxl_unittests
Change-Id: I3b02c5adef4d7983302deb4e13b6d447e4f36644