[inspect] New reader interface for inspect.
This change factors out the reading of inspect objects to a new
interface consisting of ObjectReader and ObjectHierarchy.

ObjectReader is a wrapper around the Inspect API to support easily
reading and processing the data at a specific node in the inspect tree.

ObjectHierarchy simplifies the task of recursively reading an entire
inspection hierarchy over FIDL (using ObjectReader) or directly from an
in-memory object. It supports an option to limit the depth, and it
includes the "base path" for a node, which is the path of children from
the start node for each individual node.

TEST=runtests
CF-273: #done

Change-Id: I81b56c9f5535da521ee8fbb33ecadb176ace9525