[debugger] Add the ability to hide inline frames.
This isn't quite usable but makes a smaller reviewable chunk. This API
will likely need to change according to how it's used.

It adds the ability to hide inline frames at the top of the stack. This
will be used by the stepping functions to control which virtual frame
execution is in for cases where the frame is ambiguous due to inline
frames.

TEST=unit

Change-Id: I927eec2634430efee0c51090c78ea3d3cb376612