[KMS] Basic sturcture for KMS.
Basic sturcture for KMS. Define the CryptoProvider, ProviderKey and KmsKey.
KmsKey contains a provider key generated by the provider. The generated
KmsKey is inserted into a key map to make sure there is only one
instance of a KmsKey in memory.

Write a MockProvider class to mock a real crypto provider to enable unit
testing.

Test: fx run-test kms_tests
Bug: None
Change-Id: I410aa9451b743681dc50959c4b712d7a866c55bf