[debugger] Improve thread controllers.
Enhances the "until" thread controller to be more flexible about how
the frame is compared (there is now a comparison function).

The FinishThreadController is simplified to take a stack and an index to
finish. The complexity of "from" and "to" frames are no longer
necessary since it can directly request the information from the Stack.

The test FinishThreadControllerTest.Finish is updated. Previously it did
not send a complete context (the later calls lacked the
normally-required 2-deep stack) which confused the fingerprint
computation in the new patch.

TEST=unit

Change-Id: I6a028f15743b414af47d52bc87604d67179362b5