[threads][test] Avoid scheduling flake in pthreads core test
This test needs both to not hang indefinitely, and to return quickly
on the typical expected success.

As a comprimise, poll at high (relatively to scheduler and human
perception of CQ, at least) frequency for 10 seconds, rather than once
for much longer.

This all correctly makes an int atomic, as it is accessed without
other synchronization from multiple threads.

Bug: FLK-35 #done
Test: core tests
Change-Id: I466106aae2364ae50ac3daba707d14a306d66925