[netboot] Fix netboot.Discover local address.
This change fixes the value returned by netboot.Discover("..", false).
It now returns the same address as `netaddr --local` which is the
address that routes from the remote nodename to the local machine.
This is necessary for things like registering a package server
with a target device, since amberctl needs to reach the server running
on the local machine over http.

IN-853 #comment

Change-Id: I1e75d77ed489064a4080d251333351df293cf125