[boot] Try transfering 20 times instead of 10.
We're seeing a failure to send transfer the images about 20 times per
day in CI and CQ.
This may just be due to noise on the Ethernet switches used, so it's
possible that trying again will eventually succeed.
I'll monitor the flake rate after this change and revert it if it
doesn't improve things.

Bug: IN-914 #comment
Test: None
Change-Id: I7d484b7c077d7bc8f6ddd67c9e39d1f61178a223