[sysmem] Add FIDL command to query protected memory info
sysmem-assistant in Garnet will use this to communicate with the secmem
TA to actually set the memory as protected.

For now there's a race condition where a process could allocate
protected memory that's not protected (and thus not writable by hardware
in protected mode) but that's only for a short time period on boot, when
nothing should need protected memory.

Test: vkcube --protected_output on astro
MTWN-231 #comment

Change-Id: Iabfbc0ec5d2703d6b44126db722728968f24b624