[netstack] Move eth.Controller to the link package
The link package is currently empty, and the bridge package's dependency
on the eth package makes it not possible to run the bridge tests as a
host test.

In addition, there's nothing ethernet-specific about these lifecycle
methods or states - "link.Controller" and "link.State" are better names
for these anyway.

Tested: CQ
Change-Id: I3b2d31fe8391a717d11a5525360c8e1e121c3039