[scripts] Add ramdisk option to package-image.sh
Mediatek devices crash in the bootloader when the
decompressed kernel is written to the ramdisk
region. Add an option to package-image.sh that
will prevent a ramdisk from being added to the
boot image, and use this option in the Cleo and
mt8167s_ref flash scripts. The default behavior
has not changed, so other devices shouldn't be
affected.

ZX-3356 #done
Test: mt8167 ref board boots.
Test: Ran package-image.sh and inspected the boot
      image with a hex editor.

Change-Id: If084e162d2bfe9b4ad37339c9b4daacc43791861