[kernel] Read-writer lock for handle table and arena
This replaces the mutex for the handle table lock and the per object type arena
with a rwlock. The purpose of changing to a rwlock is to allow for parallelism
in the operation of ProcessDispatcher::GetHandleLocked and Hande::FromU32, which
are both very commonly performed, short, read only operations.

ZX-3340

Tests: Ran garnet with lockdep checking enabled on vim2 + NUC.
Benchmarks on NUC + vim2 (see ZX-3340)

Change-Id: Iec967fe70437539a84a9ef523b41705883e90e49