[zircon][kernel][userboot] Prevent kcounter vmo release
VmObjectPaged::CreateFromROData moves pages from wired kernel memory
into vmos. If those vmos are closed, the pages will be freed. Handle
that situation for the kcounter vmos by either unmapping the pages from
the kernel aspace or leaking a reference to the created vmo.

Also rename CreateFromROData to CreateFromWiredPages and added a
parameter indicating whether the vmo should have exclusive access to
the pages, to try and make the api clearer.

Test: CQ, kcounter
Change-Id: Ie139fa74825ffe4191ec4754ab020110ecc9c410