[ulib][zxcrypt] Fix memory alignment/VLA problem in volume.cpp
volume.cpp currently uses a variable length array of type char
allocated on the stack to pass to block_impl_queue. This is not guaranteed
to be aligned to the required ABI requirements, and is also unsafe.

This patch moves the dynamic memory allocation to the heap
and ensures the allocation is at least aligned to alignof(block_op_t)

Test: Ensure that all automated tests pass on ARM64 and x86_64.
Change-Id: I5ccab789b02d9411b84bcd2b05c44d722852429d