[test] Add a board driver for integration tests.
This board driver enumerates a list of devices based on metadata passed
in via the ZBI and adds them.

In the future, it could be possible to provide describe fake resources
to pass to the spawned devices.

A recipricol library will be made for tests to easily construct isolated
devmgrs with the devices they would like to test enumerated.

Tested: With future changes.
Change-Id: Ifc4cb0f539d6661682c889a6fd133f4c1927e587