[dev][usb-bus] Opt out of CBs part 2 : track queued.
This adds a pending_reqs list to each endpoint.
Requests are removed from the list when a complete callback is received.

Next CL will modify this to handle the case of erroneous requests potentially
completing out of order and having to add additional callbacks.

ZX-932 #comment

TEST= ran usb tests on FX3 (usb-fwloader && runtests -t usb-test),
and tested USB audio still works.

Change-Id: I25f70875376a24327c95113ee0e6d7509efcbeec