[fidl] Prepare Zircon code for single value for ordinals.
When we remove explicit ordinals from the codebase, GenOrdinal
and Ordinal will have the same value.  You aren't allowed to
have multiple cases of the same value in switch statements.
Changing this to if statements mitigates it.

Alternative considered: If we were to remove the GenOrdinal case,
we would simply have to add it again when we do the 64-bit ordinal
migration, which would be avoidable code churn.

Test: fx full-build and pave, CQ.
Change-Id: I927b30e209395a591d43878c4391c56e3fd19d17