[devmgr] Introduce a test suite for libdriver
This is currently in Garnet due to relying on some libraries with
libstdc++ dependencies.  It would be better for this suite to be closer
to devmgr, and we should do that once the various in-flight
restructurings make it easier.

The test suite itself starts with a few basic tests.  Once this is in,
we should start expanding it.

Test: Ran these new tests overnight, saw no flakes.

Change-Id: I78f20ae2c2f4dde0eea6b7c4633d64f416cb3f31