[dev][mt-musb-peripheral] Fix breakage from previous CL.
ep->current_req was not being set, we were leaking a usb::UnownedRequest
resulting in the request completing with ZX_ERR_INTERNAL
and Endpoint::ep_num was not being initialized correctly.
As a result, the driver was completely broken.

This was not caught in testing because the bootloader was booting off
the boot_b partition instead of boot_a, meaning that I was not actually
testing my changes. The zircon flashing scripts have since been changed
to flash both partitions now, to avoid that problem.

TEST: paving garnet on cleo

Change-Id: I15a0cb144cddf09051226cd44a60a557329dd29f