[git-follow-log] prints follow logs for a path
git log --follow is limited to a single object, due to cost of performing
otherwise. This implementation does otherwise - it first computes all objects
that exist under a given path, and combines the followed history of all of
those objects.

This command takes a good amount of time for large subtrees. It is
recommended that users provide relatively specific paths for low latency
results.

Test: PATH=$PATH:$PWD/scripts/git git follow-log scripts/gce
Change-Id: I7f4200f25872cb58b918217699c15bcc237537d3