[zircon][usb-virtual-bus] Fix deadlock in usb-virtual-bus disable method
If the virtual bus is disabled while a transaction is still pending, it is possible for the system to deadlock.
This patch updates the system to release the device and host device nodes prior to removing them from devmgr
in order to allow the operation to complete, and the device to be removed successfully.

The underlying cause is the global devmgr lock that is held when processing a message in a synchronous fashion.

Test: Validate that the system no longer deadlocks when disabling the bus while a transaction is still in progress.
Change-Id: Ie37661bbeb99ea1ef1f27e0ca32d629e73620c9a