[garnet][sl4f] Add wlan connect and disconnect
Add support to sl4f to call connect and disconnect.  Connect currently
takes the ssid and a password and will attempt to connect to the
network using the first wlan interface (as returned by the list
interfaces fidl call).  Disconnect support is also added and calls
disconnect on all available wlan interfaces.

Note: some formatting changes are included as well

Bug: WLAN-907
Test: Tested using (not yet merged) WlanScanTest: act.py -tc WlanScanTest
Change-Id: I4753f15fba430d83f32c4935d49b6c5ab933e62e