[guest][integration] Add Linux virtio-net tests
The Linux networking stack does not allow for receiving raw packets, so
we must convince it that there is a real network on the other end. To
achieve this we do the bare minimum. First we send a ICMP neighbour
advertisement packet to tell the guest about the host network. Then we
use UDP to send and receive our test packets.

MAC-210 #done

TEST=guest_integration_tests

Change-Id: I5bb1aa79167ce83e03c52776b8fbfd6148693e8a