[fidlc] Support for byte and bytes alias
Adding two new built-ins, `byte` and `bytes`, to represent uint8 and
vector<uint8> respectively. The `bytes` alias can be bounded, or marked
nullable, i.e. it is a replacement to using a vector of uint8.

We'd like to support such aliasing generically, and discussions to that
effect have been ongoing, but this feels like a nice small addition
while we're waiting on a larger more general solution.

FIDL-483 #progress

Test: added one
Test: make -j 72 HOST_USE_ASAN=true tools && ./build-x64/host_tests/fidl-compiler-test
Change-Id: I10660b93478a9f4b9bb148ae3df2f17df9665073