[debugger] Launch components through zxdb.
This is a copy of I324b31d14eb9d6b8ac2643f8eb8dbb4673215deb

This is an intermediatte CL, as a lot of the details have to be ironed
out and not trivial changes need to be made to the MessageLoop in order
for this feature to work reliably.

Moved the Regex utility class to debug_ipc/helper in order to be usable
by both the client and the agent.

TEST=manual

Change-Id: I1b73ee57478c210f1d854b81a8e7a8e23bc764f5