[debugger] Report error on changed stacks.
Previously frame fingerprint computation attempted to handle the case
where a stack changed out from under it and report the new index if
possible.

This change removes the logic and reports an error. The added complexity
doesn't seem to be worth the complexity. When stepping and doing other
operations requiring fingerprints, the thread should generally be kept
stopped so the frames should not change.

TEST=unit

Change-Id: I8b0ca6cbfcf68ff67d6a495ce8a622fbb200c10e