[input_reader] Create generic mouse support
Add a HID parser for generic mice based on the
generic touchpad parser. Any mouse which is reported
as a BootMouse will keep the same hardcoded path through
input_reader.

This patch was tested using both a boot mouse and
the new paradise touchpad firmware which reports itself
as a mouse.

ZX-3351 #done

TEST: fx run-image-test input_reader_unittests
Change-Id: I6660663c0504ba42db89f68bfc5c42d82431ed30