[zircon][hid-parser][input] Use new hid-parser fields
Use the new hid parser fields for the hid.c file.
This simplifies a lot of the work that was done in
this function.

The unit tests have to be updated because the extra
8 bits for the report ID are calculated in a
different location now.

Also a bug is fixed in the hid-parser library.
A report with no fields will now have a 0 byte length
instead of a 1 byte length.

TEST: runtests -a -t hid-test

Change-Id: I3b1202bbc2ecaf98e342cdec3cbe9c3438e6436f