[dart-test] ignore relative lib imports
The dart lint rule 'avoid_relative_lib_imports'
looks for the string '/lib/' in an import statement
and fails if it finds it. This will break any test that
is in tree under a lib directory.

If a test lives in //topaz/lib/foo/example/test/foo_test.dart
the test will not run so we need to ignore the
avoid_relative_lib_imports lint rule.

TEST=None

Change-Id: I18c84f0d09ab797d2fc157ee18de8afe29c0339e