[zircon][lib][zxtest]: Fix ASSERT_BYTES_EQ
before:
ASSERT_BYTES_EQ(a,b, msg)

now:
ASSERT_BYTES_EQ(a,b, size, msg);

Note: a and b are pointer types.

ZX-3376 #done

Force pointers to void* for comparison and dumping, this prevents
nullptr_t to be overloaded into any other type, causing it to be dumped
as an integer. This is in Cpp.

TEST=zxtest-integration-test

Change-Id: Id877203de446a5cf9044b6697174294dd8aa64ce