[virtcon][display] Add FIDL method to get single framebuffer
AllocateVmo will be removed soon, replaced with sysmem, but
SimpleDisplay can only display on an existing VMO so that won't work for
it.

A new FIDL method is added that will only be used for retrieving a
single framebuffer on displays where that's the only supported method
for drawing. This is more straightforward than the existing mechanism,
which uses a hack where AllocateVmo always returns the existing VMO.

Test: virtcon on Intel, with intel-i915 and simple-display

Change-Id: Ib4df5478e2defb18895d1de5ec72d3551fb7a81f