[guest][vmm] Fix DevMem::YieldInverseRange
If the range described by |base| and |size| does not overlap with any
range, the final call to yield would be made with invalid values.

In addition, we should make sure that MemorySpecs with
MemoryPolicy::HOST_DEVICE should be skipped when generating normal
memory ranges for the guest to use.

MAC-180

Test: Added an extra test to DevMemTest.
Change-Id: I30c98f8f7131bb47c6aaf517c6d3a4eb2a922652