[products] Add top-level products directory and move 'base' to it
This adds a top-level products directory as the fuchsia-git proposal
specifies and moves the 'base' configuration definition from
//garnet/products/ to //products. This also updates 'fx set' to prefer
looking in //products and //boards before searching through the petal
repository names when users specify only a short name for a product or
board configuration.  This doesn't change the contents of base, just
moves it.

This leaves a warning (implement with a gn print statement) for the old
path to give developers time to re-run fx set or update their args if
they are driving gn manually.

Test: fx set x64 --product base
Change-Id: I44bfcbbfa3c81b911b24ee139aa69e57ac6f8376