[scenic][SCN-1054][SCN-1020] Make Scenic right handed.
This change flips the direction of Scenic's Z axis with respect to it's X and Y
axes in order to make the orientation of the vector space in which the scene
graph exists right handed.

Note that the change to SetTranslationRH is temporary and will be undone once
clients are migrated to the new behavior.

TESTED: full stack on paradise, scenic_tests, garnet/examples/ui, ermine,
tiles, flutter, mediaplayer_tests, mediaplayer_skia, media_player_test_util

Change-Id: I78ed9d3996291d8c9525df8d56b0aef3d5e1e926