[debugger] Fix crash in attaching.
The serialization of the SymbolTables request attempted to copy a
std::string by memcpy which caused corrupt memory when the message was
received in the debug_agent. This change fixes the serialization to do
it per-field.

Fixes the capture in the zxdb client for module symbol loading since
nothing guarantees the process or the module is still in scope by the
time the async message is received. These now use weak poitners.

Comments out the symbol table fetching code. This seems to reliably
crash for me.

ZX-3380 #done
TEST=manual

Change-Id: I62fb6a3fefab446c4e50768e69c39bab9c4dbf72