[dev][usb][xhci] More conversion to C++
This was previously approved and submitted as
https://fuchsia-review.googlesource.com/c/fuchsia/+/250734
but was later reverted due to a crash when mexecing or rebooting.
This new version of the CL fixes that bug.

- Use new/delete instead of malloc/calloc/free
- Use fbl::Mutex and fbl::AutoLock instead of mtx_t
- Use ddk::MmioBuffer instead of mmio_buffer_t
- Use ddk::IoBuffer instead of io_buffer_t
- Use libzx handle wrappers
- Fix some memory leaks

The xdc code is unchanged. Will work on that in a spearate pass.
More xhci clean up coming in subsequent CLs.

TEST: Manual USB testing on NUC and VIM2.
No longer crashes when suspending the driver due to mexec or reboot.

Change-Id: I6319c0c1a3e9da197566a806e408f9e45bc70ae7