[fdio] Make fdio_t an opaque object.
This patch causes all clients of the "fdio_t" object
type to access it indirectly. This improves encapsulation, and will
allows individual files to avoid a dependency on stdatomic.h,
making conversion to C++ easier.

Additionally:
- Avoid exposing |fdio_free()| directly. |fdio_t| is refcounted,
and should only be destroyed via |fdio_release()|.
- Remove the |magic| field of |fdio_t|, which was treated
as write-only, and never validated (even in debug configurations).
- Update the API for |fdio_alloc()| to require an ops table, and
ensure that all future accesses to this ops table through the fdio
object are read-only (the ops table cannot be changed on a live object).
- Avoid using |fdio_alloc()| for non-fdio objects (this was the case
in namespace.c, and meant that attempting to |fdio_free()| the object
would have caused undefined behavior.

TEST=CQ

Change-Id: Ifde4412582c57ae775766527df70230ff2274f81