[zircon][lib][zxtest]: Improve macro test coverage.
The test library relies on macros as API. Testing macros is
painful, but should prevent any unexpected behavior. We rely
on test_context_t and attribute cleanup for expectation setting
and verification. This is whether an assertion or expectation
aborted a test by mistake or a wrong comparison was made.

This should provide more trust in changing implementation in the future
if needed.

Added api for checking if the current test has any failures.

TEST=zxtest-integration-tests

Change-Id: I5db98dcb1495f3c375b5e61646aef8f7b79807d0