[input_reader] Pull code out of hid_decoder.cc
This is the first step of cleaning up input_reader.
Pull code out of hid_decoder and fdio_hid_decoder
so that the main responsibilities are reading the
report descriptor and reports in a binary array.

This patch also simplifies the mock_hid_decoder
and replaces the current mock tests with full
end-to-end tests that send binary report
descriptors and reports.

SCN-1251 #comment

TEST: Used all unit and end-to-end tests in input_reader.
      Also manually tested the HID devices on my
      paradise and frank. Also manually tested the
      devices that needed setup by spoofing their
      HID report descriptors.

Change-Id: Id053be03474e95b83e164bb39f5054e722374efa