[display][mipi-dsi] Create a mipi-dsi library and use it in astro
Currently mipi-dsi functionality is baked into the astro display driver.
Now that mediatek also has a dsi interface, it needs to live in a
separate folder.
mipi-dsi driver should really have two components: A library that
accepts some inputs and creates a valid dsi packet. The second part
would accept this packet and send it through the actual IP block that is
implementing the dsi protocol (designware, mediatek, etc).

This CL is the first part of the overall change. It create a mipi-dsi
library that can be used by any driver to create a valid dsi packet.

The logic in the mipi-dsi library is not new.

Also removed bind.c

Test: Tested on Astro
Change-Id: Iaf9cf217ca0a448b54a67fbcf1ad629e958dedb7