[fuchsia.process] Start transition to address API review
This CL starts the transition to address API review comments that
require breaking changes to fuchsia.process. In API, we decided to
remove the string error messages from Launch() and
CreateWithoutStarting(). The Launch() method is used by prebuilt
components, so we need to change the ABI with a soft transition.

Test: spawn-test
Change-Id: I9f4c1518eaa466fbe7aa141a65cad214d586bc5b