[scsilib] Implement REPORT LUNs to constrain the SCSI scan
Dramatically speeds up SCSI target scan.

The SCSI REPORT LUNs command is addressed to either LUN 0 or a
well-known LUN on a target. It returns an inventory of all addressable
LUNs on that target, which can be used to constrain a SCSI bus scan.

Use the REPORT LUNs returned size to know how many LUNs are attached to
a target and scan up to that limit; this commit doesn't use the
inventory itself, as the LUNs can be in many formats and we don't want
to have to parse each type.

If REPORT LUNs fails, assume there are no LUNs attached. This is only
valid for SCSI-3 and higher versions. (We should examine the Sense data
to look at failure reason, but ExecuteCommandSync doesn't return it in a
usable format just yet).

When running on QEMU with TCG with SCSI targets at 1:1291 and 2:100,
this finds target 2:100 in under a second, versus 8 seconds before.
The guest is interactively usable while the scan is running now, whereas
before keystrokes would take a second to appear.

ZX-2314

Tested:
Started QEMU with disks on 1:1291 and 2:100
 ./scripts/run-zircon-x64 -- \
    -drive if=none,id=hd,file=/tmp/blk.img \
    -device virtio-scsi-pci,id=scsi  \
    -device scsi-hd,drive=hd,scsi-id=1,lun=1291\
    \
    -drive if=none,id=hd2,file=/tmp/blk2.img \
    -device scsi-hd,drive=hd2,scsi-id=2,lun=100

Saw:
[00003.057] 02914.03299> 1:1291 QEMU     QEMU HARDDISK
[00003.057] 02914.03299> 511 blocks of 512 bytes
[00004.033] 02914.03299> 2:100 QEMU     QEMU HARDDISK
[00004.033] 02914.03299> 511 blocks of 512 bytes

Change-Id: I0403cdc4d2e33b85d55c5c7ba6ea4e3ca3c974b0